:hammer: Move around commands and providers
What was the commands directory becomes the providers directory (the
providers of commands for the command palette). What was (mostly) the
messages directory gets turned into the commands directory; this is now
where command messages are defined.

Those messages that aren't "command messages" remain where they were.
