+++ src/braindrop/app/commands/__init__.py
@@ -1,14 +1,66 @@
-"""Provides the command palette command provides for the application."""
+"""Provides command-oriented messages for the application.
+
+These messages differ a little from other messages in that they have a
+common base class and provide information such as help text, binding
+information, etc.
+"""
 
 ##############################################################################
 # Local imports.
-from .collections import CollectionCommands
-from .commands_provider import CommandsProvider
-from .main import MainCommands
-from .tags import TagCommands
+from .base import Command
+from .commands import (
+    AddRaindrop,
+    ChangeTheme,
+    CheckTheWaybackMachine,
+    ClearFilters,
+    CompactMode,
+    CopyLinkToClipboard,
+    DeleteRaindrop,
+    Details,
+    EditRaindrop,
+    Escape,
+    Help,
+    Logout,
+    Quit,
+    Redownload,
+    Search,
+    SearchCollections,
+    SearchTags,
+    ShowAll,
+    ShowUnsorted,
+    ShowUntagged,
+    TagOrder,
+    VisitLink,
+    VisitRaindrop,
+)
 
 ##############################################################################
 # Exports.
-__all__ = ["CollectionCommands", "CommandsProvider", "MainCommands", "TagCommands"]
+__all__ = [
+    "Command",
+    "AddRaindrop",
+    "ChangeTheme",
+    "CheckTheWaybackMachine",
+    "ClearFilters",
+    "CompactMode",
+    "CopyLinkToClipboard",
+    "DeleteRaindrop",
+    "Details",
+    "EditRaindrop",
+    "Escape",
+    "Help",
+    "Logout",
+    "Quit",
+    "Redownload",
+    "Search",
+    "SearchCollections",
+    "SearchTags",
+    "ShowAll",
+    "ShowUnsorted",
+    "ShowUntagged",
+    "TagOrder",
+    "VisitLink",
+    "VisitRaindrop",
+]
 
 ### __init__.py ends here
