+++ src/braindrop/app/providers/__init__.py
@@ -1,0 +1,14 @@
+"""Provides the command palette command provides for the application."""
+
+##############################################################################
+# Local imports.
+from .collections import CollectionCommands
+from .commands_provider import CommandsProvider
+from .main import MainCommands
+from .tags import TagCommands
+
+##############################################################################
+# Exports.
+__all__ = ["CollectionCommands", "CommandsProvider", "MainCommands", "TagCommands"]
+
+### __init__.py ends here
