+++ src/braindrop/app/widgets/navigation.py
@@ -24,8 +24,9 @@ from textual.widgets.option_list import Option
 ##############################################################################
 # Local imports.
 from ...raindrop import API, Collection, SpecialCollection, Tag, TagData
+from ..commands import ShowAll, ShowUnsorted, ShowUntagged
 from ..data import LocalData, Raindrops
-from ..messages import ShowAll, ShowCollection, ShowTagged, ShowUnsorted, ShowUntagged
+from ..messages import ShowCollection, ShowTagged
 from .extended_option_list import OptionListEx
